project3_requirements (1).txt
Flask==3.0.0
flask-cors==4.0.0
torch==2.1.2
transformers==4.35.2
Pillow==10.1.0
Werkzeug==3.0.1
ftfy==6.1.3
regex==2023.12.25